+++ ZuluContent/Engines/Chat/ChatUser.cs
@@ -52,21 +52,11 @@ namespace Server.Engines.Chat
             return false;
         }
 
-        public void SendMessage(int number, string param1 = null, string param2 = null)
-        {
-            if (Mobile.NetState != null)
-            {
-                Mobile.Send(new ChatMessagePacket(Mobile, number, param1, param2));
-            }
-        }
+        public void SendMessage(int number, string param1 = null, string param2 = null) =>
+            SendMessage(number, Mobile, param1, param2);
 
-        public void SendMessage(int number, Mobile from, string param1, string param2)
-        {
-            if (Mobile.NetState != null)
-            {
-                Mobile.Send(new ChatMessagePacket(from, number, param1, param2));
-            }
-        }
+        public void SendMessage(int number, Mobile from, string param1 = null, string param2 = null) =>
+            Mobile.NetState.SendChatMessage(from.Language, number, param1, param2);
 
         public bool IsIgnored(ChatUser check) => Ignored.Contains(check);
