+++ ZuluContent/Engines/Chat/Packets.cs
@@ -4,7 +4,7 @@ namespace Server.Engines.Chat
 {
     public sealed class ChatMessagePacket : Packet
     {
-        public ChatMessagePacket(Mobile who, int number, string param1, string param2) : base(0xB2)
+        public ChatMessagePacket(string lang, int number, string param1, string param2) : base(0xB2)
         {
             param1 ??= string.Empty;
             param2 ??= string.Empty;
@@ -13,9 +13,9 @@ namespace Server.Engines.Chat
 
             Stream.Write((ushort)(number - 20));
 
-            if (who != null)
+            if (lang != null)
             {
-                Stream.WriteAsciiFixed(who.Language, 4);
+                Stream.WriteAsciiFixed(lang, 4);
             }
             else
             {
