+++ ZuluContent/Items/Books/BookContent.cs
