+++ ZuluContent/Items/Misc/BulletinEquip.cs
@@ -1,0 +1,14 @@
+namespace Server.Items
+{
+    public struct BulletinEquip
+    {
+        public int itemID;
+        public int hue;
+
+        public BulletinEquip(int itemID, int hue)
+        {
+            this.itemID = itemID;
+            this.hue = hue;
+        }
+    }
+}
