+++ ZuluContent/Items/Misc/OilCloth.cs
@@ -4,9 +4,9 @@ namespace Server.Items
 {
     public class OilCloth : Item, IScissorable, IDyable
     {
-        public override int LabelNumber { get; } = 1041498; // oil cloth
+        public override int LabelNumber => 1041498; // oil cloth
 
-        public override double DefaultWeight { get; } = 1.0;
+        public override double DefaultWeight => 1.0;
 
 
         [Constructible]
