+++ ZuluContent/Items/Weapons/SpearsAndForks/Pitchfork.cs
@@ -3,19 +3,19 @@ namespace Server.Items
     [FlipableAttribute(0xE87, 0xE88)]
     public class Pitchfork : BaseSpear
     {
-        public override int DefaultStrengthReq { get; } = 15;
+        public override int DefaultStrengthReq => 15;
 
-        public override int DefaultMinDamage { get; } = 4;
+        public override int DefaultMinDamage => 4;
 
-        public override int DefaultMaxDamage { get; } = 16;
+        public override int DefaultMaxDamage => 16;
 
-        public override int DefaultSpeed { get; } = 45;
+        public override int DefaultSpeed => 45;
 
-        public override int InitMinHits { get; } = 31;
+        public override int InitMinHits => 31;
 
-        public override int InitMaxHits { get; } = 60;
+        public override int InitMaxHits => 60;
 
-        public override int DefaultMaxRange { get; } = 2;
+        public override int DefaultMaxRange => 2;
 
         [Constructible]
         public Pitchfork() : base(0xE87)
