+++ ZuluContent/ZuluContent.csproj
@@ -34,7 +34,7 @@
 
   <PropertyGroup>
     <PostBuildEvent Condition="'$(OS)' != 'UNIX'">
-      robocopy "$(SolutionDir)Distribution\*" "$(OutDir)\..\" /FFT /XA:H
+      robocopy "$(SolutionDir)Distribution\Configuration" "$(OutDir)\..\Configuration" /FFT /XA:H
       robocopy "$(SolutionDir)Distribution\Data" "$(OutDir)\..\Data" /MIR /FFT /XA:H
     </PostBuildEvent>
   </PropertyGroup>
